Add files via upload
Hi all, 

This is my project. 

CODE: https://pastebin.com/BknQRKa8 

Goals: Create an interactive virtual audio system that reacts to music and gives feedback based on how you play with the controller. The overall design is in the form of an orb. 

Credits: 

https://www.youtube.com/watch?v=jKnWxlQMy6o
https://cycling74.com/tutorials/making-connections-connecting-a-joystick-to-maxmspjitter
https://docs.cycling74.com/max5/tutorials/max-tut/communicationschapter01.html
Ml Star / ADSR in Package Manager 

Intro: 

This is an interactive visualizer that works to make the most of whatever music your listening to. Simply play any song and your computer mic will pic of the audio data and pulse the ORB to match the musics vibe. 


But, it gets better.

Simply attach any gaming controller and the visualizer turns into a mini home theater to match your audiophile home speaker set up. You can now play with all the controls and jam out to any song with taps on keys and by twiddling the joy sticks. 

Each part of the controller messes with a different part of the set up. 

Meanwhile, there is full customization in terms of the visualizer. You can keep it a slick orb or you can turn it into a massive mess of colors. Simply use the square buttons to tweak to your hearts desire. There are also options to mess with how big the orb is and how reactive the orb is to the music. 

All in all, reality becomes one with your music as your hands become part of such an enticing visualizer. 

Enjoy!